requirements.txt
aws-cdk-lib==2.43.1
constructs==10.1.111
cdk-nag==2.18.15